rename file old and create a new one
